+++ ProgramV2/src/bldc/bldc.c
@@ -8,9 +8,9 @@ Timer dt_timer;
 
 Timer speed_dt_timer;
 
-static inline double BLDC_GetEncoder(SensoredVectorControl* svc, uint16_t encoder_val, double encoder_offset_theta) {
+static inline float BLDC_GetEncoder(SensoredVectorControl* svc, uint16_t encoder_val, float encoder_offset_theta) {
   uint16_t correction_adc_val = encoder_val * svc->adc_correction_factor;
-  double theta = correction_adc_val * ADC2RADIAN;          // 0〜2πの範囲に変換
+  float theta = correction_adc_val * ADC2RADIAN;           // 0〜2πの範囲に変換
   theta = NormalizeRadians(theta - encoder_offset_theta);  // オフセット値を引いて正規化
 
   // ローパスフィルタ
@@ -29,23 +29,23 @@ static inline double BLDC_GetEncoder(SensoredVectorControl* svc, uint16_t encode
   }
 }
 
-static inline double BLDC_GetMaxEncoderVal(uint16_t encoder_val) {
+static inline float BLDC_GetMaxEncoderVal(uint16_t encoder_val) {
   static uint16_t max_val = 4000;
   if (encoder_val > max_val) max_val = encoder_val;
   return max_val;
 }
 
-static inline double BLDC_GetSpeed(double theta, double dt) {
-  static double pre_speed = 0;
-  static double pre_theta = 0;
-  static double pre_delta_theta = 0;
+static inline float BLDC_GetSpeed(float theta, float dt) {
+  static float pre_speed = 0;
+  static float pre_theta = 0;
+  static float pre_delta_theta = 0;
 
   if (pre_theta == theta) {
     pre_theta = theta + pre_delta_theta;
     return pre_speed;
   }
 
-  double delta_theta = theta - pre_theta;
+  float delta_theta = theta - pre_theta;
 
   // 0と2πの境目を跨いだ場合の補正
   if (delta_theta > PI) delta_theta -= TWO_PI;
@@ -55,7 +55,7 @@ static inline double BLDC_GetSpeed(double theta, double dt) {
   if (Abs(delta_theta) > PI) delta_theta = pre_delta_theta;
   pre_delta_theta = delta_theta;
 
-  double speed = delta_theta / dt;
+  float speed = delta_theta / dt;
   speed = speed * (1 - SPEED_LPF) + pre_speed * SPEED_LPF;
   pre_speed = speed;
   pre_theta = theta;
@@ -63,9 +63,9 @@ static inline double BLDC_GetSpeed(double theta, double dt) {
   return speed;
 }
 
-static inline double BLDC_PIDControl(PIDController* pid, double error, double dt) {
+static inline float BLDC_PIDControl(PIDController* pid, float error, float dt) {
   // 比例項
-  double p_term = pid->kp * error;
+  float p_term = pid->kp * error;
 
   // 積分項
   pid->integral += pid->ki * error * dt;
@@ -73,11 +73,11 @@ static inline double BLDC_PIDControl(PIDController* pid, double error, double dt
   if (pid->integral < -pid->output_limit) pid->integral = -pid->output_limit;
 
   // 微分項
-  double d_term = pid->kd * (error - pid->prev_error) / dt;
+  float d_term = pid->kd * (error - pid->prev_error) / dt;
   pid->prev_error = error;
 
   // 出力の計算
-  double output = p_term + pid->integral + d_term;
+  float output = p_term + pid->integral + d_term;
   if (output > pid->output_limit) output = pid->output_limit;
   if (output < -pid->output_limit) output = -pid->output_limit;
 
@@ -93,12 +93,12 @@ static inline void BLDC_SetEncoder(SensoredVectorControl* svc, uint16_t* encoder
     BLDC_OpenLoopDrive(0.15, phase);
     HAL_Delay(1);
   }
-  svc->adc_correction_factor = (double)MAX_ADC_VAL / svc->max_encoder_val;
+  svc->adc_correction_factor = (float)MAX_ADC_VAL / svc->max_encoder_val;
 
   // 電気角度を0にオフセットする
-  double offset_sum = 0;
+  float offset_sum = 0;
   for (uint8_t i = 0; i < svc->pole_pairs; i++) {
-    double theta_sum = 0;
+    float theta_sum = 0;
     BLDC_OpenLoopDrive(0.2, 0);
     HAL_Delay(200);
     BLDC_OpenLoopDrive(0.4, 0);
@@ -124,7 +124,7 @@ static inline void BLDC_SetEncoder(SensoredVectorControl* svc, uint16_t* encoder
   Flash_WriteData(FLASH_USER_START_ADDR, &write_data, sizeof(write_data));
 }
 
-static inline void BLDC_WritePwm(double u, double v, double w) {
+static inline void BLDC_WritePwm(float u, float v, float w) {
   u = Constrain(u, MIN_DUTY, MAX_DUTY);
   v = Constrain(v, MIN_DUTY, MAX_DUTY);
   w = Constrain(w, MIN_DUTY, MAX_DUTY);
@@ -157,7 +157,7 @@ void BLDC_Init(SensoredVectorControl* svc, bool do_set_encoder, uint16_t* encode
   printf("(From Flash)max_encoder_val: %ld, encoder_offset_theta: %.6f\n", read_data.max_encoder_val, read_data.encoder_offset_theta);
   svc->max_encoder_val = read_data.max_encoder_val;            // エンコーダーの最大値
   svc->encoder_offset_theta = read_data.encoder_offset_theta;  // エンコーダーのオフセット値
-  svc->adc_correction_factor = (double)MAX_ADC_VAL / svc->max_encoder_val;
+  svc->adc_correction_factor = (float)MAX_ADC_VAL / svc->max_encoder_val;
 
   // PIDコントローラ
   // 速度制御
@@ -181,17 +181,17 @@ void BLDC_Stop(bool brake) {
   }
 }
 
-void BLDC_OpenLoopDrive(double amp, double phase) {
+void BLDC_OpenLoopDrive(float amp, float phase) {
   phase = NormalizeRadians(phase);
 
-  double u = 0.5 + 0.5 * amp * Cos(phase);
-  double v = 0.5 + 0.5 * amp * Cos(phase - TWO_THIRDS_PI);
-  double w = 0.5 + 0.5 * amp * Cos(phase + TWO_THIRDS_PI);
+  float u = 0.5 + 0.5 * amp * Cos(phase);
+  float v = 0.5 + 0.5 * amp * Cos(phase - TWO_THIRDS_PI);
+  float w = 0.5 + 0.5 * amp * Cos(phase + TWO_THIRDS_PI);
 
   BLDC_WritePwm(u, v, w);
 }
 
-void BLDC_SensoredVectorControlDrive(SensoredVectorControl* svc, uint16_t encoder_value, double supply_volt) {
+void BLDC_SensoredVectorControlDrive(SensoredVectorControl* svc, uint16_t encoder_value, float supply_volt) {
   // 制御周期の取得
   svc->dt = Timer_Read(&dt_timer);
   Timer_Reset(&dt_timer);
@@ -213,21 +213,21 @@ void BLDC_SensoredVectorControlDrive(SensoredVectorControl* svc, uint16_t encode
   svc->amp = Constrain(svc->amp, -1, 1);
 
   // 正弦波を生成
-  double u, v, w;
+  float u, v, w;
   u = 0.5 + 0.5 * svc->amp * Sin(svc->elec_theta);
   v = 0.5 + 0.5 * svc->amp * Sin(svc->elec_theta - TWO_THIRDS_PI);
   w = 0.5 + 0.5 * svc->amp * Sin(svc->elec_theta + TWO_THIRDS_PI);
   BLDC_WritePwm(u, v, w);
 }
 
-void BLDC_SpeedControl(SensoredVectorControl* svc, double target_speed) {
+void BLDC_SpeedControl(SensoredVectorControl* svc, float target_speed) {
   // 最大速度制限
   if (target_speed > MAX_SPEED) target_speed = MAX_SPEED;
   if (target_speed < -MAX_SPEED) target_speed = -MAX_SPEED;
 
   // 最大加速度制限
-  static double prev_target_speed = 0;
-  double accel = (target_speed - prev_target_speed) / svc->dt;
+  static float prev_target_speed = 0;
+  float accel = (target_speed - prev_target_speed) / svc->dt;
   if (accel > MAX_ACCEL) accel = MAX_ACCEL;
   if (accel < -MAX_ACCEL) accel = -MAX_ACCEL;
   target_speed = prev_target_speed + accel * svc->dt;
@@ -236,9 +236,9 @@ void BLDC_SpeedControl(SensoredVectorControl* svc, double target_speed) {
   svc->amp_volt = -BLDC_PIDControl(&svc->speed_pid, target_speed - svc->speed, svc->dt);
 }
 
-void BLDC_PositionControl(SensoredVectorControl* svc, double target_position) {
+void BLDC_PositionControl(SensoredVectorControl* svc, float target_position) {
   // 位置制御のためのPID計算
-  double error = target_position - (svc->mech_theta);  // 目標位置と現在位置の誤差
+  float error = target_position - (svc->mech_theta);  // 目標位置と現在位置の誤差
 
   // 0と2πのまたぎ対策
   while (error > PI) error -= TWO_PI;
@@ -246,6 +246,6 @@ void BLDC_PositionControl(SensoredVectorControl* svc, double target_position) {
   svc->amp_volt = -BLDC_PIDControl(&svc->position_pid, error, svc->dt);
 }
 
-void BLDC_TorqueControl(SensoredVectorControl* svc, double target_torque) {
+void BLDC_TorqueControl(SensoredVectorControl* svc, float target_torque) {
   svc->amp_volt = target_torque;
 }
